Move path normalization under core
Place the shared path normalization helper under core/pathmatch instead of a top-level internal package so the helper lives with the model it supports while still avoiding core/analysisutil import cycles.

Constraint: Keep analysisutil.NormalizeMatchPath public behavior unchanged

Confidence: high

Scope-risk: narrow

Tested: go test -count=1 ./core ./core/analysisutil ./rules/types; git diff --check

Not-tested: full test suite
