+++ core/analysisutil/path.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/NamhaeSusan/go-arch-guard/core"
-	"github.com/NamhaeSusan/go-arch-guard/internal/pathmatch"
+	"github.com/NamhaeSusan/go-arch-guard/core/pathmatch"
 	"golang.org/x/tools/go/packages"
 )
